SCreate separate utility function for handling image compression and saving
