+++ client/src/services/tools.ts
@@ -1,3 +1,4 @@
+import imageCompression from "browser-image-compression";
 
 export function BreakDownTime(standard_time: String): Array<String> {
     const monthMap = new Map([
@@ -66,4 +67,31 @@ export function IsFutureDate(standard_time: String): boolean {
     }
 
     return true;
+}
+
+// For now, store as a base-64 string
+// In the future, we would like to upload this to an external object/file storage instead
+export function CompressAndSaveImg(input: string, name: string, saveHandler: (img: string) => void): void {
+    let reader = new FileReader();
+    reader.onload = async () => {
+        if(reader.result) {
+            let res: string = reader.result.toString();
+            saveHandler(res);
+        }
+    };
+
+    const options = {
+        maxSizeMB: 1,
+        maxWidthOrHeight: 250,
+        useWebWorker: true
+    };
+
+    fetch(input)
+        .then(res => res.blob())
+        .then(blob => {
+        const file = new File([blob], name, { type: "image/jpg" });
+        imageCompression(file, options)
+            .then((data: File) => reader.readAsDataURL(data))
+            .catch((error: Error) => console.log(error.message));
+    });
 }
